+++ src/main/kotlin/com/nylas/resources/Auth.kt
@@ -165,7 +165,7 @@ class Auth(private val client: NylasClient) {
       .adapter(UrlForAuthenticationConfig::class.java)
       .toJson(config)
     JsonHelper.jsonToMap(json).forEach { (key, value) ->
-      if(key == "scope") {
+      if (key == "scope") {
         url.addQueryParameter("scope", config.scope?.joinToString(separator = " "))
       } else {
         url.addQueryParameter(key, value.toString())